Set Timeout element in http.Client{}
An `http.Client{}` will default to never timing out; in the event of a server hang the client will never finish without a `Timeout` set.

In order to provide the user with more granularity I've adjusted the `Request` function parameters to accept an `*http.Client` object. This allows users to set things like `net.Dialer.Timeout`, `http.Transport.TLSHandshakeTimeout`, `http.Transport.IdleConnTimeout`, etc... prior to passing an `*http.Client` to submit their request. In the event that `nil` is passed the function will define `client` with a `Timeout` of 15 seconds.

It is worth noting that this change will break apps using the previous definition of `Request`.